Look up the ES cluster config from Secrets Manager, not env vars
