+++ .sbt_metadata/items.json
@@ -3,6 +3,7 @@
   "folder" : "items",
   "dependencyIds" : [
     "stacks",
-    "search_common"
+    "search_common",
+    "config"
   ]
 }
