+++ .sbt_metadata/search.json
@@ -3,6 +3,7 @@
   "folder" : "search",
   "dependencyIds" : [
     "display",
-    "search_common"
+    "search_common",
+    "config"
   ]
 }
