+++ common/config/src/main/scala/weco/catalogue/config/ElasticConfig.scala
@@ -1,0 +1,23 @@
+package weco.catalogue.config
+
+import com.sksamuel.elastic4s.Index
+
+case class ElasticConfig(
+  worksIndex: Index,
+  imagesIndex: Index
+)
+
+trait ElasticConfigBase {
+
+  // We use this to share config across API applications
+  // i.e. The API and the snapshot generator.
+  val pipelineDate = "2022-02-03"
+}
+
+object PipelineClusterElasticConfig extends ElasticConfigBase {
+  def apply(): ElasticConfig =
+    ElasticConfig(
+      worksIndex = Index(s"works-indexed-$pipelineDate"),
+      imagesIndex = Index(s"images-indexed-$pipelineDate")
+    )
+}
