+++ common/config/src/main/scala/weco/catalogue/config/PipelineElasticClientBuilder.scala
@@ -1,12 +1,11 @@
-package weco.api.snapshot_generator.config.builders
+package weco.catalogue.config
 
 import com.sksamuel.elastic4s.ElasticClient
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient
 import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest
-import weco.catalogue.display_model.PipelineClusterElasticConfig
 import weco.elasticsearch.ElasticClientBuilder
 
-object SnapshotGeneratorElasticClientBuilder {
+object PipelineElasticClientBuilder {
 
   // We create a new pipeline cluster for every pipeline, and we want the
   // snapshot generator to read from that cluster.
@@ -16,7 +15,7 @@ object SnapshotGeneratorElasticClientBuilder {
   // we let the snapshot generator decide which set of secrets to read, which
   // in turn sets which cluster it reads from.
 
-  def apply(): ElasticClient = {
+  def apply(name: String): ElasticClient = {
     implicit val secretsClient: SecretsManagerClient =
       SecretsManagerClient.builder().build()
 
@@ -32,10 +31,10 @@ object SnapshotGeneratorElasticClientBuilder {
       s"elasticsearch/pipeline_storage_$pipelineDate/protocol"
     )
     val username = getSecretString(
-      s"elasticsearch/pipeline_storage_$pipelineDate/snapshot_generator/es_username"
+      s"elasticsearch/pipeline_storage_$pipelineDate/$name/es_username"
     )
     val password = getSecretString(
-      s"elasticsearch/pipeline_storage_$pipelineDate/snapshot_generator/es_password"
+      s"elasticsearch/pipeline_storage_$pipelineDate/$name/es_password"
     )
 
     ElasticClientBuilder.create(
@@ -48,8 +47,8 @@ object SnapshotGeneratorElasticClientBuilder {
   }
 
   private def getSecretString(
-    id: String
-  )(implicit secretsClient: SecretsManagerClient) = {
+                               id: String
+                             )(implicit secretsClient: SecretsManagerClient) = {
     val request =
       GetSecretValueRequest
         .builder()
