+++ items/src/main/resources/application.conf
@@ -4,12 +4,6 @@ apm.secret=${?apm_secret}
 apm.service.name=${?apm_service_name}
 apm.environment=${?apm_environment}
 
-es.host=${?es_host}
-es.port=${?es_port}
-es.username=${?es_username}
-es.password=${?es_password}
-es.protocol=${?es_protocol}
-
 http.host="0.0.0.0"
 http.port=${?app_port}
 http.externalBaseURL=${?app_base_url}
