+++ requests/src/main/resources/application.conf
@@ -7,12 +7,6 @@ apm.secret=${?apm_secret}
 apm.service.name=${?apm_service_name}
 apm.environment=${?apm_environment}
 
-es.host=${?es_host}
-es.port=${?es_port}
-es.username=${?es_username}
-es.password=${?es_password}
-es.protocol=${?es_protocol}
-
 aws.metrics.namespace=${?metrics_namespace}
 sierra.api.key=${?sierra_api_key}
 sierra.api.secret=${?sierra_api_secret}
