+++ search/src/main/scala/weco/api/search/Main.scala
@@ -2,13 +2,15 @@ package weco.api.search
 
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
-import weco.elasticsearch.typesafe.ElasticBuilder
 import weco.Tracing
 import weco.api.search.models.{ApiConfig, CheckModel, QueryConfig}
 import weco.api.search.swagger.SwaggerDocs
+import weco.catalogue.config.{
+  PipelineClusterElasticConfig,
+  PipelineElasticClientBuilder
+}
 import weco.typesafe.WellcomeTypesafeApp
 import weco.typesafe.config.builders.AkkaBuilder
-import weco.catalogue.display_model.ApiClusterElasticConfig
 import weco.http.WellcomeHttpApp
 import weco.http.monitoring.HttpMetrics
 import weco.http.typesafe.HTTPServerBuilder
@@ -28,8 +30,8 @@ object Main extends WellcomeTypesafeApp {
 
     implicit val apiConfig: ApiConfig = ApiConfig.build(config)
 
-    val elasticClient = ElasticBuilder.buildElasticClient(config)
-    val elasticConfig = ApiClusterElasticConfig()
+    val elasticClient = PipelineElasticClientBuilder(name = "catalogue_api")
+    val elasticConfig = PipelineClusterElasticConfig()
 
     CheckModel.checkModel(elasticConfig.worksIndex.name)(elasticClient)
     CheckModel.checkModel(elasticConfig.imagesIndex.name)(elasticClient)
