+++ snapshots/snapshot_generator/src/main/resources/application.conf
@@ -1,9 +1,3 @@
 aws.sns.topic.arn=${?topic_arn}
 aws.sqs.queue.url=${?queue_url}
 aws.metrics.namespace=${?metrics_namespace}
-es.host=${?es_host}
-es.port=${?es_port}
-es.username=${?es_username}
-es.password=${?es_password}
-es.protocol=${?es_protocol}
-es.bulk-size=${?es_bulk_size}
