+++ snapshots/snapshot_generator/src/main/scala/weco/api/snapshot_generator/Main.scala
@@ -4,13 +4,15 @@ import akka.actor.ActorSystem
 import com.amazonaws.services.s3.AmazonS3
 import com.sksamuel.elastic4s.ElasticClient
 import com.typesafe.config.Config
-import weco.api.snapshot_generator.config.builders.SnapshotGeneratorElasticClientBuilder
 import weco.api.snapshot_generator.models.SnapshotGeneratorConfig
 import weco.api.snapshot_generator.services.{
   SnapshotGeneratorWorkerService,
   SnapshotService
 }
-import weco.catalogue.display_model.PipelineClusterElasticConfig
+import weco.catalogue.config.{
+  PipelineClusterElasticConfig,
+  PipelineElasticClientBuilder
+}
 import weco.messaging.sns.NotificationMessage
 import weco.messaging.typesafe.{SNSBuilder, SQSBuilder}
 import weco.storage.typesafe.S3Builder
@@ -34,7 +36,7 @@ object Main extends WellcomeTypesafeApp {
     )
 
     implicit val elasticClient: ElasticClient =
-      SnapshotGeneratorElasticClientBuilder()
+      PipelineElasticClientBuilder(name = "snapshot_generator")
 
     implicit val s3Client: AmazonS3 = S3Builder.buildS3Client
